Add model and mock content for locations
This required creating the necessary resource descriptions, adding the
migrations service, and adding specific migrations for the content
which needs to appear on the website.  Do note that the content of
these migrations is mock content and will be updated before the site
goes live.  However, the created URIs for delivery-kinds should stay
constant from now on.