Add basket resource properties for order status
The payment status was described as a property on the basket in the
domain.lisp.  This property should be changed only by backend services
which is an exercise for later.  Current changes include the basket
status, indicating whether the basket is still being edited or not, as
well as knowing when the basket was last changed.

There's WIP to verify the amount payed on the basket and check that
the user did not update the basket's contents between ordering and
paying.  We should execute a check and let the alarms go off when this
happens so we can best support users.