Add properties for ordering baskets
The location now has an email property so we can extract that
information for ordered baskets.  This is not the cleanest approach but
it works for now.

The basket also gets a connected user, which we can use to identify the
person.  The person gets a specific email address which we can use to
contact the user regardless of the method through which they log
in (should more login methods be added in the future).