Change predicate used for plu
The plu was using adms:identifier.  That's great, but we could use something
else so we can use adms:identifier or other external systems too with a more
complex object behind it.  We could also use adms:identifier with a complex
object for the current PLU, but current indexing systems and access rights don't
allow us to specify such complex paths yet.