Add Veeakker as a BusinessEntity
This entity is added to the LFW graph so the description and such can easily be
updated.  We would ideally have this in the public graph (because it's not
lfw-import's responsibility) but that currently doesn't allow us to update
properties (they might be written to an incorrect graph).

A possible solution is to only let the lfw-graph contain the lfw specific links
to their identifiers and to have organizations live in another graph.